:recycle: #12 エラーメッセージの表示部分は単独で動作確認出来るように別Componentに分離
