+++ src/components/ErrorFallback/ErrorFallback.tsx
@@ -1,14 +1,7 @@
 import type { FC } from 'react';
-import { Alert, Button } from '@mantine/core';
+import { Button } from '@mantine/core';
 import type { FallbackProps } from 'react-error-boundary';
-
-const createDisplayErrorMessage = (error: Error) => {
-  if (error.name === 'GitHubAccountNotFoundError') {
-    return 'GitHubアカウントは見つかりませんでした。';
-  }
-
-  return '予期せぬエラーが発生しました。申し訳ありませんが、しばらく時間が経ってからお試し下さい。';
-};
+import { ErrorMessage } from '@/components/ErrorFallback/ErrorMessage';
 
 export const ErrorFallback: FC<FallbackProps> = ({
   error,
@@ -16,9 +9,7 @@ export const ErrorFallback: FC<FallbackProps> = ({
 }) => {
   return (
     <>
-      <Alert title="エラーが発生しました。" color="red">
-        {createDisplayErrorMessage(error)}
-      </Alert>
+      <ErrorMessage error={error} />
       <Button onClick={resetErrorBoundary}>もう一度試す</Button>
     </>
   );
